refactor: code duplication fixes
